refactor: folder for all conversation-related components
