+++ src/core/components/Conversation/Content.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect, memo } from 'react';
-import { sanitizeContent } from '../utils/sanitize';
+import { sanitizeContent } from '../../utils/sanitize';
 import styles from './Content.module.css';
-import { useAppContext } from '../context/useAppContext';
+import { useAppContext } from '../../context/useAppContext';
 
 export interface ContentProps {
   content: string;
