+++ src/core/components/Conversation/StreamingText.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen, act } from '@testing-library/react';
 import { StreamingText } from './StreamingText';
-import { TextStreamStore } from '../stores/textStreamStore';
+import { TextStreamStore } from '../../stores/textStreamStore';
 
 describe('StreamingText', () => {
   it('renders with the initial snapshot of the store', () => {
