+++ src/core/components/Conversation/StreamingTextContent.tsx
