+++ src/core/components/Conversation/ThinkingContent.module.css
